extract proxies for testing
