+++ src/test/java/org/javacv/detect/face/haar/DetectorServiceTest.java
@@ -25,7 +25,6 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.verify;
-import static org.junit.jupiter.api.Assertions.assertFalse;
 
 /**
  * Unit test for {@link DetectorService}.
